Reimplement tooltips using portals (PR #1301)
* Reimplement tooltips using portals (Fixes #794)

* Reduce number of elements used for mounting tooltips

Also use the common root-overlay element as a mount point.

Fixes #794, fixes #1253
